Make save method return generated key

+++ src/main/java/ru/ruscalworld/storagelib/Storage.java
@@ -9,7 +9,7 @@ public interface Storage {
     <T> T find(@NotNull Class<T> clazz, String key, Object value) throws Exception;
     <T> T retrieve(Class<T> clazz, int id) throws Exception;
     <T> List<T> retrieveAll(Class<T> clazz) throws Exception;
-    <T extends DefaultModel> int save(T model) throws Exception;
+    <T extends DefaultModel> long save(T model) throws Exception;
     ConverterProvider getConverterProvider();
     void registerConverter(Class<?> clazz, Converter<?> converter);
     void actualizeStorageSchema() throws Exception;
